Primeiro desafio nodejs

+++ www/ignite/nodejs/desafio-01/src/routes.js
@@ -6,34 +6,43 @@ const database = new Database()
 
 export const routes = [ 
   {
+  method: 'POST',
+  path: buildRoutePath ('tasks'),
+  handler:(req, res) => {
+    const { title, description } = req.body
+    
+    if (!title) {
+      return res.writeHead(400).end(
+        JSON.stringify({ massage: 'Descrição Requerida'})
+      )
+      }
+
+      const task = {
+        id: randomUUID(),
+        title,
+        description,
+        completed_at: null,
+        created_at: new Date(),
+        updated_at: new Date(),
+      }
+
+      database.insert('tasks', task)
+
+      return res.writeHead(201).end()
+    }
+    },
+  {
     method: 'GET',
     path: buildRoutePath ('/tasks'),
     handler:(req, res) => {
       const { search } = req.query
 
-      const users = database.select('tasks', search ? {
-        name: search,
-        email: search,
-      } : null)
+      const tasks = database.select('tasks', {
+        title: search,
+        description: search,
+      })
 
       return res.end(JSON.stringify(tasks))
-    }
-  },
-  {
-    method: 'POST',
-    path: buildRoutePath ('tasks'),
-    handler:(req, res) => {
-      const { name, email } = req.body
-
-      const task = {
-        id: randomUUID(),
-        name,
-        email,       
-      }
-
-      database.insert('tasks', task)
-      
-      return res.writeHead(201).end()
     }
   },
   {
@@ -41,11 +50,24 @@ export const routes = [
     path: buildRoutePath('(/tasks/):id'),
     handler: (req, res) => {
       const { id } = req.params
-      const { name, email } = req.body
+      const { title, description } = req.body
 
-      database.update('users', id, {
-        name,
-        email,
+      if(!title && !description) {
+        return res.writeHead(400).end(
+          JSON.stringify({ massage: 'titulo ou descrição foram requiridos'})
+        )
+      }
+
+      const [task] = database.select('task', {id})
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
+
+      database.update('tasks', id, {
+        title: title ?? task.title,
+        description: description ?? task.description,
+        updated_at: new Date()
       })
 
       return res.writeHead(204).end()
@@ -57,9 +79,35 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params
 
+      const [task] = database.select('tasks', {id})
+
+      if(!task) {
+        return res.writeHead(404).end()
+      }
+
       database.delete('tasks', id)
 
       return res.writeHead(204).end()
-    },
+    }
+  },
+  {
+    method: 'PATH',
+    path: buildRoutePath('/tasks/:id/complete'),
+    handler: (req, res) => {
+      const { id } = req.params
+      
+      const [task] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
+
+      const isTaskCompleted = !!task.created_at
+      const completed_at = isTaskCompleted ? null : new Date()
+
+      database.update('tasks', id { completed_at })
+
+      return res.writeHead(204).end()
+    }
   }
 ]
